Extract TestResponseBase.
